Перенес экшен selectedOffer из слайса offers в offer и актулизировал тесты
