+++ src/store/offers/offers-slices.ts
@@ -1,4 +1,4 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { createSlice } from '@reduxjs/toolkit';
 import { TOffer } from '../../types/offers.ts';
 import { fetchOffers } from '../api-actions/offers-action.ts';
 import { NameSpace } from '../../constants/name-space.ts';
@@ -6,23 +6,17 @@ import { NameSpace } from '../../constants/name-space.ts';
 type TOffersState = {
   offers: TOffer[];
   isLoading: boolean;
-  selectedOffer: TOffer['id'] | null;
 }
 
 const initialState: TOffersState = {
   offers: [],
   isLoading: false,
-  selectedOffer: null
 };
 
 const offersSlices = createSlice({
   name: NameSpace.Offers,
   initialState,
-  reducers: {
-    addSelectedOffer(state, action: PayloadAction<TOffer['id'] | null>) {
-      state.selectedOffer = action.payload;
-    }
-  },
+  reducers: {},
   extraReducers(builder) {
     builder
       .addCase(fetchOffers.pending, (state) => {
@@ -36,6 +30,4 @@ const offersSlices = createSlice({
 });
 
 export default offersSlices.reducer;
-
-export const { addSelectedOffer } = offersSlices.actions;
